GVM's (internal) cpu affinities should not affect spawned processes
When configured with enable-multiple-threaded-vms the runtime will pin
the VM threads to certain CPUs.  On POSIX and Windows this setting
will be inherited be created/forked processes, which is usually not
wanted.

This patch removes any affinity settings (that is sets up the affinity
to all CPUs) for a newly created process.  Currently this only works
for Windows and POSIX but not MacOS.

Tested on linux only.