GVM's (internal) cpu affinities should not affect spawned processes
When configured with enable-multiple-threaded-vms the runtime will pin
the VM threads to certain CPUs.  At least on linux the affinity of the
thread which called exec will be inherited, which is usually unwanted.

On Windows only the process-affinity-mask will be "inherited", but as
we only modify the thread-affinity mask, everything works as desired.

I don't know anything about Mac-OS.

This patch removes any affinity settings (that is sets up the affinity
to all CPUs) for a newly created process on linux.  Tested and seems
to work.